Word to opencyc

+++ agent/CycReasoningAgent.py
@@ -2,20 +2,20 @@ import logging
 import os
 import time
 
-from agent.CycLServerAgent import CycLServerAgent
+from service.CycLService import CycLService
 from service.OllamaService import OllamaService
 
 ollama_model = "gemma3n:latest"
 
 
 class CycReasoningAgent:
-    def __init__(self, ollama_service: OllamaService, cyc_server_agent: CycLServerAgent, logger: logging.Logger):
+    def __init__(self, ollama_service: OllamaService, cyc_server_agent: CycLService, logger: logging.Logger):
         self.logger = logger
         self.cyc_server_agent = cyc_server_agent
         self.ollama_service = ollama_service
         self.microtheories = []
         self.predicate_microtheories = []
-        self._populate_microtheories()
+        # self._populate_microtheories()
         self.predicates = None
 
     def _populate_microtheories(self):
@@ -38,18 +38,24 @@ class CycReasoningAgent:
                        "Given a microtheory name, I will provide a concise and accurate answer describing the microtheory name."
                        "The given microtheory is: ")
         for mt in self.microtheories:
-            start_time = time.time()
             mt_dir = os.path.join(os.getcwd(), "runtime", "microtheories", mt)
             if not os.path.exists(mt_dir):
                 os.makedirs(mt_dir)
             filename = os.path.join(mt_dir, f"{mt}.gemma-analysis.txt")
+            if os.path.exists(filename):
+                self.logger.info(f"Skipping {mt} due to pre-existing file.")
+                continue
+            start_time = time.time()
             result = self.ollama_service.one_shot(ollama_model, mt, base_prompt)
+            elapsed_time = time.time() - start_time
             if result is None:
                 result = "None"
-            elapsed_time = time.time() - start_time
-            with open(filename, 'w') as fh:
-                fh.write(result)
-                fh.write(f"-------\n")
-                fh.write(f"Inference for microtheory took {elapsed_time} seconds.")
-            fh.close()
+            self.logger.info(f"Inference for microtheory took {elapsed_time} seconds.")
+            try:
+                with open(filename, 'w') as fh:
+                    fh.write(result)
+                    fh.close()
+            except Exception as fe:
+                self.logger.error(f"Exception {fe} caught while processing {str(result)}")
+                continue
